change txsRate from uint to float64

+++ network/p2p/peer.go
@@ -433,7 +433,7 @@ type Peer struct {
 
 	id        string
 	version   uint
-	txsRate   uint
+	txsRate   float64
 	bandwidth float64
 	address   common.Address
 
@@ -492,14 +492,14 @@ func (p *Peer) SetHead(hash common.Hash, td *big.Int) {
 	p.td.Set(td)
 }
 
-func (p *Peer) TxsRate() uint {
+func (p *Peer) TxsRate() float64 {
 	p.lock.RLock()
 	defer p.lock.RUnlock()
 
 	return p.txsRate
 }
 
-func (p *Peer) SetTxsRate(txs uint) {
+func (p *Peer) SetTxsRate(txs float64) {
 	p.lock.Lock()
 	defer p.lock.Unlock()
 	p.txsRate = txs
